Refactor crypto and hash packages
