+++ backend/internal/oauth/jwks/service.go
@@ -30,8 +30,8 @@ import (
 	"crypto/sha1" //nolint:gosec
 
 	"github.com/asgardeo/thunder/internal/system/config"
-	"github.com/asgardeo/thunder/internal/system/crypto/hash"
 	"github.com/asgardeo/thunder/internal/system/error/serviceerror"
+	"github.com/asgardeo/thunder/internal/system/hash"
 )
 
 // JWKSServiceInterface defines the interface for JWKS service.
