+++ backend/internal/system/hash/model.go
